sort.php
This is a PHP script that takes the "starred.json" file as an input (see http://www.nick-web.co.uk/ee/archives/64 and http://www.nick-web.co.uk/ee/archives/31 for more information).

This file DOES NOT need script.sh to work. Just take this file, put it in a php-cabable directory with the "starred.json" file, and call the script via a browser.

Save the resultant HTML file as instapaper-export.html and use that to import your starred items to Pocket.